+++ packages/api-proxy/src/v3/eth/contract-calls/contract-calls.controller.ts
@@ -18,7 +18,6 @@ import {
   ApiBadRequestResponse,
   ApiExtraModels,
   ApiOperation,
-  ApiResponse,
   ApiTags,
 } from "@nestjs/swagger";
 import {