+++ packages/api-proxy/src/v3/eth/transactions/transactions.controller.ts
@@ -30,7 +30,6 @@ import {
   ApiResponseContentGenerator,
   AuthErrorResponses,
   AuthHeaders,
-  PaginationResponse,
   PathParams,
   Queries,
   ReadMeExtension,