+++ packages/api-proxy/src/v3/eth/transfers/transfers.service.ts
@@ -6,19 +6,27 @@ import { PaginationDTO } from "../dto/pagination.dto";
 import { Pagination } from "@haechi-labs/henesis-wallet-core/lib/types";
 import { EthValueTransferEvent } from "@haechi-labs/henesis-wallet-core/lib/events";
 import { object } from "../../../utils/object";
+import { getPaginationMeta } from "../../../utils/pagination";
 
 @Injectable()
 export class TransfersService {
   public async getTransfers(
     sdk: SDK,
-    option: GetTransfersOption
+    options: GetTransfersOption,
+    path: string
   ): Promise<PaginationDTO<TransferDTO>> {
     const result: Pagination<EthValueTransferEvent> =
       await sdk.eth.events.getValueTransferEvents(
-        object(GetTransfersOption.toSDKOption(option))
+        object(GetTransfersOption.toSDKOption(options))
       );
     return {
-      pagination: result.pagination as any,
+      pagination: getPaginationMeta(
+        path,
+        options.page,
+        options.size,
+        result.pagination.totalCount,
+        options
+      ),
       results: result.results.map(TransferDTO.fromValueTransferEvent),
     };
   }