+++ packages/api-proxy/src/v3/eth/wallets/dto/send-coin-request.dto.ts
@@ -1,38 +1,52 @@
-import { ApiPropertyOptional } from "@nestjs/swagger";
-import { ApiModelPropertyOptional } from "@nestjs/swagger/dist/decorators/api-model-property.decorator";
+import {
+  ApiModelProperty,
+  ApiModelPropertyOptional,
+} from "@nestjs/swagger/dist/decorators/api-model-property.decorator";
+
+export const EXAMPLE_ETHEREUM_SEND_COIN_REQUEST_DTO: SendCoinRequestDTO = {
+  ticker: "ETH",
+  to: "0xab28d146e860e0b132695c941f706d783a158345",
+  amount: "100000000",
+  passphrase: "passphrase",
+  gasPrice: "8000000000",
+  metadata: "metadata",
+};
 
 export class SendCoinRequestDTO {
-  /**
-   * 암호화폐의 기호 (ticker)
-   * @example ETH
-   */
+  @ApiModelProperty({
+    description: "암호화폐의 기호 (ticker)",
+    example: EXAMPLE_ETHEREUM_SEND_COIN_REQUEST_DTO.ticker,
+  })
   ticker: string;
-  /**
-   * 암호화폐의 받을 지갑 주소
-   * @example 0xab28d146e860e0b132695c941f706d783a158345
-   */
+
+  @ApiModelProperty({
+    description: "암호화폐를 받을 지갑 주소",
+    example: EXAMPLE_ETHEREUM_SEND_COIN_REQUEST_DTO.to,
+  })
   to: string;
-  /**
-   * 암호화폐의 양 (단위: wei, peb)
-   * @example 100000000
-   */
+
+  @ApiModelProperty({
+    description: "암호화폐의 양 (단위: wei, peb)",
+    example: EXAMPLE_ETHEREUM_SEND_COIN_REQUEST_DTO.amount,
+  })
   amount: string;
-  /**
-   * 지갑의 비밀번호
-   * @example passphrase
-   */
+
+  @ApiModelProperty({
+    description: "지갑의 비밀번호",
+    example: EXAMPLE_ETHEREUM_SEND_COIN_REQUEST_DTO.passphrase,
+  })
   passphrase: string;
-  /**
-   * 트랜잭션에 사용할 gas price (단위: wei) null일 경우, Henesis가 자동으로 5분 안에 채굴될 수 있는 값으로 설정합니다.
-   * @example: 8000000000
-   */
-  @ApiPropertyOptional()
-  gasPrice: number;
+
+  @ApiModelPropertyOptional({
+    description:
+      "트랜잭션에 사용할 gas price (단위: wei) null일 경우, Henesis가 자동으로 5분 안에 채굴될 수 있는 값으로 설정합니다.",
+    example: EXAMPLE_ETHEREUM_SEND_COIN_REQUEST_DTO.gasPrice,
+  })
+  gasPrice: string;
 
   @ApiModelPropertyOptional({
     description: "기타 정보 기록용 메타 데이터 (255자 제한)",
     example: "metadata",
-    default: "metadata",
   })
   metadata?: string;
 }