+++ packages/api-proxy/src/v3/fil/wallets/wallets.service.ts
@@ -20,6 +20,7 @@ import { PaginationDTO } from "../dto/pagination.dto";
 import { FilDepositAddress } from "@haechi-labs/henesis-wallet-core/lib/fil/depositAddress";
 import { MasterWalletBalanceDto } from "../dto/master-wallet-balance.dto";
 import { PaginationOptionsDTO } from "../dto/pagination-options.dto";
+import { getPaginationMeta } from "../../../utils/pagination";
 
 @Injectable()
 export class WalletsService {
@@ -77,7 +78,8 @@ export class WalletsService {
       new BN(request.amount),
       request.passphrase,
       null,
-      request.gasPremium ? new BN(request.gasPremium) : null
+      request.gasPremium != null ? new BN(request.gasPremium) : null,
+      request.metadata
     );
     return TransferDTO.fromTransfer(transfer);
   }
@@ -93,7 +95,8 @@ export class WalletsService {
     const flush: FilFlush = await masterWallet.flush(
       request.targets,
       request.passphrase,
-      request.gasPremium ? new BN(request.gasPremium) : null
+      request.gasPremium != null ? new BN(request.gasPremium) : null,
+      request.metadata
     );
     return FlushDTO.fromFlush(flush);
   }
@@ -101,14 +104,21 @@ export class WalletsService {
   public async getDepositAddresses(
     sdk: SDK,
     masterWalletId: string,
-    options: GetDepositAddressesOptionsDTO
+    options: GetDepositAddressesOptionsDTO,
+    path: string
   ): Promise<PaginationDTO<DepositAddressDTO>> {
     const masterWallet: FilMasterWallet = await sdk.fil.wallets.getMasterWallet(
       masterWalletId
     );
     const data = await masterWallet.getDepositAddresses(options);
     return {
-      pagination: data.pagination,
+      pagination: getPaginationMeta(
+        path,
+        options.page,
+        options.size,
+        data.pagination.totalCount,
+        options
+      ),
       results: data.results.map(DepositAddressDTO.fromDepositAddress),
     };
   }
@@ -122,7 +132,7 @@ export class WalletsService {
       masterWalletId
     );
     return DepositAddressDTO.fromDepositAddress(
-      await masterWallet.createDepositAddress(request.name, request.passphrase)
+      await masterWallet.createDepositAddress(request.name)
     );
   }