+++ packages/core/src/fil/wallet.ts
@@ -157,13 +157,11 @@ export class FilMasterWallet extends FilAbstractWallet {
 
   async createDepositAddress(
     name: string,
-    passphrase?: string,
     otpCode?: string
   ): Promise<FilDepositAddress> {
     const wallet = await this.client.get<MasterWalletDTO>(this.baseUrl);
-    const depositAddressKey = this.keychains.derive(
+    const depositAddressKey = this.keychains.deriveFromPublicKey(
       this.getAccountKey(),
-      passphrase,
       wallet.nextChildNumber
     );
     const depositAddressData = await this.client.post(
@@ -227,7 +225,8 @@ export class FilMasterWallet extends FilAbstractWallet {
     amount: BN,
     passphrase: string,
     otpCode?: string,
-    gasPremium?: BN
+    gasPremium?: BN,
+    metadata?: string
   ): Promise<FilTransfer> {
     const rawTransaction = await this.client.post<
       NoUndefinedField<RawTransactionDTO>
@@ -250,6 +249,7 @@ export class FilMasterWallet extends FilAbstractWallet {
           convertSignedTransactionToRawSignedTransactionDTO(signedTransaction),
         gasPremium: BNConverter.bnToHexStringOrElseNull(gasPremium),
         otpCode: otpCode,
+        metadata: metadata,
       }
     );
     return convertDtoToTransfer(transferData);
@@ -258,7 +258,8 @@ export class FilMasterWallet extends FilAbstractWallet {
   async flush(
     targets: Array<string>,
     passphrase: string,
-    gasPremium?: BN
+    gasPremium?: BN,
+    metadata?: string
   ): Promise<FilFlush> {
     const rawFlushData = await this.client.post<NoUndefinedField<RawFlushDTO>>(
       `${this.baseUrl}/flushes/build`,
@@ -276,6 +277,7 @@ export class FilMasterWallet extends FilAbstractWallet {
       `${this.baseUrl}/flushes`,
       {
         targets: flushTargets.map(convertFilFlushTargetToDto),
+        metadata: metadata,
       }
     );
     return convertDtoToFlush(flushData);